CSRF exempt on results/
Reset search was 403 crashing on live because of a CSRF error. After being unable to shed light on this error (the CSRF was passed correctly and it should work), I took the decision to CSRF exempt results/ since nothing security critical happens in this view, it seemed like an overkill.